backend/requirements.txt
black
flake8
Flask-RESTful
Flask-HTTPAuth
